requirements.txt
streamlit
requests
google-api-python-client
groq
google-generativeai 